Remove over-engineered queryset setup
Having a queryset class that late-injects an extra query starts to ruin chainable nature of things.  Ultimately, either a clone needs to get created in order to add the filter and then the method calls need to be secretly forwarded to it, or the underlying ``query`` attribute gets modified in place.  Both options are awful.