Revamp request args retrieval
``getRequestInfo()`` now does all the hard work to translate a set of contextual arguments to the proper request characteristics.  No XHR object exists yet, but GET payloads are serialized and appended to the url, POST payloads are passed through with the active collector token injected, and headers are provided as an object, containing the collector's declared Content-Type and the active X-CSRFToken.