Revamp XHR request prep
Because an XHR object can't set a POST payload without also immediately sending that payload to the server, the previous ``getRequest()`` method fell short of being able to provide a usable request.  ``prepareRequest()`` takes the return result of a call to ``getRequestInfo()``, preparing the xhr request accordingly.  A function is returned for the calling code to trigger the final sending process when ready.