Attempt to declare nodejs exports
If the execution setting isn't nodejs, then this is irrelevant and shouldn't trigger an error.  I'd like to arrange to have this as its own explicit file that calls in the ES5 stuff, but then I get right back to this dilemma of needing the ES5 version to export its variable!

Note that exporting the DjangoInputCollection constructor function directly resulted in weirdness. I might have been doing it wrong, though, so we'll revisit this.