Add resolution of the CollectionRequest to all endpoints
This allows all endpoints in both list/detail modes to obtain a reference to a collector for a given CollectionRequest.

By default it will search for a query param called 'request' for the id, but if there is a specific object with a direction model relationship to the instance, return that instead.  It is presently unsupported to NOT provide the fallback 'request' query param when foreignkey lookups are unavailable.